+++ reapy/core/item/source.py
@@ -1,7 +1,9 @@
+import reapy
 from reapy import reascript_api as RPR
+from reapy.core import ReapyObject
 
 
-class Source:
+class Source(ReapyObject):
 
     def __init__(self, id):
         self.id = id
@@ -31,6 +33,23 @@ class Source:
         """
         _, filename, _ = RPR.GetMediaSourceFileName(self.id, "", 10**5)
         return filename
+
+    @reapy.inside_reaper()
+    @property
+    def has_valid_id(self):
+        """
+        Whether ReaScript ID is still valid.
+
+        For instance, if source has been deleted, ID will not be valid
+        anymore.
+
+        :type: bool
+        """
+        pointer, name = self._get_pointer_and_name()
+        return any(
+            RPR.ValidatePtr2(project.id, pointer, name)
+            for project in reapy.get_projects()
+        )
 
     def length(self, unit="seconds"):
         """